CIV-904 refactor class names and variable names
