+++ src/main/common/utils/dateConverter.ts
@@ -3,11 +3,12 @@ export class DateConverter {
    * Converts the string fields entered in the Date fields to a dd-mm-yyyy Date format if the fields entered correspond to a valid Date
    */
   static convertToDate(year: string, month: string, day: string) {
-    const dob = new Date(year + '-' + month + '-' + day);
-    const dobDay = Number(day);
-    if ((dob.getDate() == dobDay)) {
-      return dob;
+    const date = new Date(year + '-' + month + '-' + day);
+    const givenDay = Number(day);
+    if ((date.getDate() == givenDay)) {
+      return date;
     }
     return null;
+
   }
 }
